Move iterative CompanyAuth test into its own file
